Revert type changes
